show hostnames for loglines involving remote logationd peers
plus some refactorings
* use la_address_t instead of char for address in
  * log_trigger(), trigger_manual_command, is_local_address(),
    create_manual_command_from_template()
  * also in lots of functions in messages.c
* compile_address_list_port_domainname() optionally looks up domainname
* some refactoring with sync_all_entries() - now takes address_t * as argument
* replaced some int types with size_t or ssize_t
* created fifo_address type to identify messages that came in via fifo

+++ src/crypto.c
@@ -22,6 +22,7 @@
 #include <assert.h>
 #include <string.h>
 #include <stdbool.h>
+#include <stddef.h>
 
 #include "ndebug.h"
 #include "crypto.h"
@@ -45,7 +46,7 @@ static unsigned char send_salt[crypto_pwhash_SALTBYTES];
  */
 
 static bool
-generate_key(unsigned char *const key, const int key_len,
+generate_key(unsigned char *const key, const size_t key_len,
                 const char *const password, const unsigned char *const salt)
 {
 	assert(key); assert(key_len > 0); assert(password); assert(salt);
@@ -151,7 +152,7 @@ pad(char *buffer, const size_t msg_len)
         assert(MSG_LEN - msg_len < 256);
 
         const unsigned char pad = MSG_LEN - msg_len;
-        for (int i=msg_len; i<MSG_LEN; i++)
+        for (size_t i=msg_len; i<MSG_LEN; i++)
                 buffer[MSG_IDX+i] = pad;
 }
